4聊天chat/doubao/chat_app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncmy==0.2.8  # 异步MySQL驱动
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.4.2
typing-extensions==4.8.0
